+++ python/neuromation/client/__init__.py
@@ -4,6 +4,7 @@ from typing import Union, Type, Optional
 import aiohttp
 from yarl import URL
 
+from neuromation.client.utils import create_registry_url
 from .abc import AbstractProgress, AbstractSpinner
 from .api import (
     API,
@@ -13,6 +14,7 @@ from .api import (
     AuthError,
     AuthenticationError,
     AuthorizationError,
+    DEFAULT_TIMEOUT,
 )
 from .jobs import (
     Jobs,
@@ -58,8 +60,6 @@ __all__ = (
     "AbstractSpinner",
 )
 
-DEFAULT_TIMEOUT = aiohttp.ClientTimeout(None, None, 30, 30)
-
 
 class Client:
     def __init__(
@@ -67,13 +67,17 @@ class Client:
         url: Union[URL, str],
         token: str,
         *,
+        registry_url: str = "",  # default value is always overwritten
         timeout: aiohttp.ClientTimeout = DEFAULT_TIMEOUT,
     ) -> None:
         if isinstance(url, str):
             url = URL(url)
         self._url = url
+        # this is temporary until we implement getting server configuration dynamically:
+        registry_url = registry_url or create_registry_url(str(self._url))
+        self._registry_url = URL(registry_url)
         assert token
-        self._config = Config(url, token)
+        self._config = Config(url, self._registry_url, token)
         self._api = API(url, token, timeout)
         self._jobs = Jobs(self._api, token)
         self._models = Models(self._api)