+++ python/tests/cli/test_rc.py
@@ -1,18 +1,21 @@
 import logging
 from pathlib import Path
 from textwrap import dedent
+from unittest import mock
 
 import pkg_resources
 import pytest
+from aiohttp import web
 from jose import jwt
 from yarl import URL
 
 from neuromation.cli import rc
-from neuromation.cli.rc import AuthToken, Config
+from neuromation.cli.login import AuthConfig, ServerConfig
+from neuromation.cli.rc import AuthToken, Config, load, save
 from neuromation.client.users import JWT_IDENTITY_CLAIM_OPTIONS
 
 
-DEFAULTS = rc.Config(url="https://platform.dev.neuromation.io/api/v1")
+DEFAULTS = rc.Config(url="https://dev.ai.neuromation.io/api/v1")
 
 
 @pytest.fixture
@@ -28,23 +31,63 @@ def patch_home_for_test(monkeypatch, nmrc):
     monkeypatch.setattr(Path, "home", home)
 
 
-def test_create(nmrc):
+def test_create__with_defaults(nmrc):
     conf = rc.create(nmrc, Config())
     assert conf == DEFAULTS
     assert nmrc.exists()
     expected_text = dedent(
         """\
-    auth_config:
-      audience: https://platform.dev.neuromation.io
-      auth_url: https://dev-neuromation.auth0.com/authorize
-      client_id: V7Jz87W9lhIlo0MyD0O6dufBvcXwM4DR
-      success_redirect_url: https://platform.neuromation.io
-      token_url: https://dev-neuromation.auth0.com/oauth/token
     github_rsa_path: ''
     pypi:
       check_timestamp: 0
       pypi_version: 0.0.0
-    url: https://platform.dev.neuromation.io/api/v1
+    registry_url: ''
+    url: https://dev.ai.neuromation.io/api/v1
+    """
+    )
+    assert nmrc.read_text() == expected_text
+
+
+def test_create__filled(nmrc):
+    config = Config(
+        url="https://dev.ai/api/v1",
+        registry_url="https://registry-dev.ai/api/v1",
+        auth_config=AuthConfig(
+            auth_url=URL("url"),
+            token_url=URL("url"),
+            client_id="client_id",
+            audience="audience",
+            callback_urls=(URL("url1"), URL("url2")),
+            success_redirect_url=URL("url"),
+        ),
+        auth_token=AuthToken(
+            token="token", expiration_time=100_500, refresh_token="refresh_token"
+        ),
+    )
+    created_config = rc.create(nmrc, config)
+    assert created_config == config
+    assert nmrc.exists()
+    expected_text = dedent(
+        """\
+    auth_config:
+      audience: audience
+      auth_url: url
+      callback_urls:
+      - url1
+      - url2
+      client_id: client_id
+      success_redirect_url: url
+      token_url: url
+    auth_token:
+      expiration_time: 100500
+      refresh_token: refresh_token
+      token: token
+    github_rsa_path: ''
+    pypi:
+      check_timestamp: 0
+      pypi_version: 0.0.0
+    registry_url: https://registry-dev.ai/api/v1
+    url: https://dev.ai/api/v1
     """
     )
     assert nmrc.read_text() == expected_text
@@ -58,62 +101,6 @@ class TestFactoryMethods:
         rc.ConfigFactory._update_config(url="http://abc.def", auth_token=auth_token)
         config2: Config = rc.ConfigFactory.load()
         assert config == config2
-
-    def test_factory_update_url(self):
-        auth_token = AuthToken.create_non_expiring("token1")
-        config: Config = Config(url="http://abc.def", auth_token=auth_token)
-        rc.ConfigFactory.update_api_url(url="http://abc.def")
-        config2: Config = rc.ConfigFactory.load()
-        assert config.url == config2.url
-
-    def test_factory_update_url_malformed(self):
-        auth_token = AuthToken.create_non_expiring("token1")
-        config: Config = Config(url="http://abc.def", auth_token=auth_token)
-        with pytest.raises(ValueError):
-            rc.ConfigFactory.update_api_url(url="ftp://abc.def")
-        config2: Config = rc.ConfigFactory.load()
-        assert config.url != config2.url
-
-    def test_factory_update_url_malformed_trailing_slash(self):
-        auth_token = AuthToken.create_non_expiring("token1")
-        config: Config = Config(url="http://abc.def", auth_token=auth_token)
-        with pytest.raises(ValueError):
-            rc.ConfigFactory.update_api_url(url="http://abc.def/")
-        config2: Config = rc.ConfigFactory.load()
-        assert config.url != config2.url
-
-    def test_factory_update_url_malformed_with_fragment(self):
-        auth_token = AuthToken.create_non_expiring("token1")
-        config: Config = Config(url="http://abc.def", auth_token=auth_token)
-        with pytest.raises(ValueError):
-            rc.ConfigFactory.update_api_url(url="http://abc.def?blabla")
-        config2: Config = rc.ConfigFactory.load()
-        assert config.url != config2.url
-
-    def test_factory_update_url_malformed_with_anchor(self):
-        auth_token = AuthToken.create_non_expiring("token1")
-        config: Config = Config(url="http://abc.def", auth_token=auth_token)
-        with pytest.raises(ValueError):
-            rc.ConfigFactory.update_api_url(url="http://abc.def#ping")
-        config2: Config = rc.ConfigFactory.load()
-        assert config.url != config2.url
-
-    def test_factory_update_url_and_auth_config(self):
-        config = rc.ConfigFactory.load()
-        assert config.url == "https://platform.dev.neuromation.io/api/v1"
-        assert config.auth_config.auth_url == URL(
-            "https://dev-neuromation.auth0.com/authorize"
-        )
-
-        rc.ConfigFactory.update_api_url(
-            url="https://platform.staging.neuromation.io/api/v1"
-        )
-
-        config = rc.ConfigFactory.load()
-        assert config.url == "https://platform.staging.neuromation.io/api/v1"
-        assert config.auth_config.auth_url == URL(
-            "https://staging-neuromation.auth0.com/authorize"
-        )
 
     def test_factory_update_id_rsa(self):
         config: Config = Config(
@@ -154,30 +141,137 @@ class TestFactoryMethods:
         jwt_hdr = """eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"""
         jwt_claims = """eyJpZGVudGl0eSI6Im1lIn0"""
         jwt_sig = """mhRDoWlNw5J2cAU6LZCVlM20oRF64MtIfzquso2eAqU"""
-        test_token = AuthToken.create_non_expiring(f"{jwt_hdr}.{jwt_claims}.{jwt_sig}")
-        config: Config = Config(
+        token = f"{jwt_hdr}.{jwt_claims}.{jwt_sig}"
+
+        rc.ConfigFactory.update_auth_token(token)
+        expected_config = Config(
             url=DEFAULTS.url,
-            auth_token=test_token,
+            auth_token=AuthToken(
+                token=token, expiration_time=mock.ANY, refresh_token=""
+            ),
             github_rsa_path=DEFAULTS.github_rsa_path,
         )
-        rc.ConfigFactory.update_auth_token(test_token.token)
-        config2: Config = rc.ConfigFactory.load()
-        assert config == config2
+        config: Config = rc.ConfigFactory.load()
+        assert config == expected_config
+
         rc.ConfigFactory.forget_auth_token()
         config3: Config = rc.ConfigFactory.load()
-        default_config: config = Config()
+        default_config: Config = Config()
         assert config3 == default_config
+
+    @pytest.fixture
+    def server_config(self):
+        return ServerConfig(
+            registry_url=URL("registry_url"),
+            auth_config=AuthConfig(
+                auth_url=URL("https://auth0"),
+                token_url=URL("https://token"),
+                client_id="this_is_client_id",
+                audience="https://audience",
+                callback_urls=(
+                    URL("https://0.0.0.0:123"),
+                    URL("https://0.0.0.0:456"),
+                    URL("https://0.0.0.0:789"),
+                ),
+                success_redirect_url=URL("https://success_redirect"),
+            ),
+        )
+
+    @pytest.fixture
+    @pytest.mark.asyncio
+    async def server_config_url(self, server_config, aiohttp_server):
+        registry_url: URL = server_config.registry_url
+        auth_config: AuthConfig = server_config.auth_config
+        JSON = {
+            "registry_url": str(registry_url),
+            "auth_url": str(auth_config.auth_url),
+            "token_url": str(auth_config.token_url),
+            "client_id": auth_config.client_id,
+            "audience": auth_config.audience,
+            "callback_urls": [str(u) for u in auth_config.callback_urls],
+            "success_redirect_url": str(auth_config.success_redirect_url),
+        }
+
+        async def handler(request):
+            return web.json_response(JSON)
+
+        app = web.Application()
+        app.router.add_get("/config", handler)
+        srv = await aiohttp_server(app)
+
+        yield srv.make_url("")
+
+        await app.cleanup()
+
+    async def test_factory_update_url(self, server_config_url, server_config, nmrc):
+        uninit_auth_config = AuthConfig.create_uninitialized()
+        uninit_config = Config()
+        auth_token = AuthToken.create_non_expiring("token1")
+
+        config = Config(url="http://url", auth_token=auth_token, github_rsa_path="path")
+        save(nmrc, config)
+        await rc.ConfigFactory.update_api_url(url=str(server_config_url))
+        config2 = load(nmrc)
+
+        assert config.url == "http://url"
+        assert config.registry_url == ""
+        assert config.auth_config == uninit_auth_config
+        assert config.auth_token == auth_token
+        assert config.github_rsa_path == "path"
+        assert config.pypi == uninit_config.pypi
+        assert config.color == uninit_config.color
+        assert config.tty == uninit_config.tty
+        assert config.terminal_size == uninit_config.terminal_size
+
+        assert config2.url == str(server_config_url)
+        assert config2.registry_url == str(server_config.registry_url)
+        assert config2.auth_config == server_config.auth_config
+        assert config2.auth_token == auth_token
+        assert config2.github_rsa_path == "path"
+        assert config2.pypi == uninit_config.pypi
+        assert config2.color == uninit_config.color
+        assert config2.tty == uninit_config.tty
+        assert config2.terminal_size == uninit_config.terminal_size
+
+    async def test_factory_update_url_malformed(self):
+        auth_token = AuthToken.create_non_expiring("token1")
+        config: Config = Config(url="http://abc.def", auth_token=auth_token)
+        with pytest.raises(ValueError):
+            await rc.ConfigFactory.update_api_url(url="ftp://abc.def")
+        config2: Config = rc.ConfigFactory.load()
+        assert config.url != config2.url
+
+    async def test_factory_update_url_malformed_trailing_slash(self):
+        auth_token = AuthToken.create_non_expiring("token1")
+        config: Config = Config(url="http://abc.def", auth_token=auth_token)
+        with pytest.raises(ValueError):
+            await rc.ConfigFactory.update_api_url(url="http://abc.def/")
+        config2: Config = rc.ConfigFactory.load()
+        assert config.url != config2.url
+
+    async def test_factory_update_url_malformed_with_fragment(self):
+        auth_token = AuthToken.create_non_expiring("token1")
+        config: Config = Config(url="http://abc.def", auth_token=auth_token)
+        with pytest.raises(ValueError):
+            await rc.ConfigFactory.update_api_url(url="http://abc.def?blabla")
+        config2: Config = rc.ConfigFactory.load()
+        assert config.url != config2.url
+
+    async def test_factory_update_url_malformed_with_anchor(self):
+        auth_token = AuthToken.create_non_expiring("token1")
+        config: Config = Config(url="http://abc.def", auth_token=auth_token)
+        with pytest.raises(ValueError):
+            await rc.ConfigFactory.update_api_url(url="http://abc.def#ping")
+        config2: Config = rc.ConfigFactory.load()
+        assert config.url != config2.url
 
 
 def test_docker_url():
-    assert DEFAULTS.docker_registry_url() == URL("https://registry.dev.neuromation.io")
-    custom_staging = rc.Config(url="https://platform.staging.neuromation.io/api/v1")
-    url = custom_staging.docker_registry_url()
-    assert url == URL("https://registry.staging.neuromation.io")
-
-    prod = rc.Config(url="https://platform.neuromation.io/api/v1")
-    url = prod.docker_registry_url()
-    assert url == URL("https://registry.neuromation.io")
+    assert DEFAULTS.registry_url == ""
+    custom_staging = rc.Config(
+        url="https://platform.io/api/v1", registry_url="https://registry.io"
+    )
+    assert custom_staging.registry_url == "https://registry.io"
 
 
 @pytest.mark.parametrize("identity_claim", JWT_IDENTITY_CLAIM_OPTIONS)
@@ -224,12 +318,13 @@ def test_create_existing(nmrc):
 def test_load(nmrc):
     document = """
         url: 'http://a.b/c'
+        registry_url: 'http://registry.a.b/c'
     """
     nmrc.write_text(document)
     nmrc.chmod(0o600)
 
     config = rc.load(nmrc)
-    assert config == rc.Config(url="http://a.b/c")
+    assert config == rc.Config(url="http://a.b/c", registry_url="http://registry.a.b/c")
 
 
 def test_load_missing(nmrc):