+++ python/tests/e2e/conftest.py
@@ -11,18 +11,17 @@ from uuid import uuid4 as uuid
 import pytest
 
 from neuromation.cli import main
-from tests.e2e.utils import (
-    FILE_SIZE_B,
-    RC_TEXT,
-    format_list,
-    format_list_pattern,
-    hash_hex,
-)
+from neuromation.client import FileStatusType
+from tests.e2e.utils import FILE_SIZE_B, RC_TEXT, hash_hex, output_to_files
 
 
 log = logging.getLogger(__name__)
 
-job_id_pattern = r"Job ID:\s*(\S+)"
+job_id_pattern = re.compile(
+    # pattern for UUID v4 taken here: https://stackoverflow.com/a/38191078
+    r"(job-[0-9a-f]{8}-[0-9a-f]{4}-[4][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12})",
+    re.IGNORECASE,
+)
 
 
 class TestRetriesExceeded(Exception):
@@ -132,7 +131,7 @@ def run(monkeypatch, capfd, tmp_path):
             out = post_out[pre_out_size:]
             err = post_err[pre_err_size:]
             if arguments[0:2] in (["job", "submit"], ["model", "train"]):
-                match = re.search(job_id_pattern, out)
+                match = job_id_pattern.match(out)
                 if match:
                     executed_jobs_list.append(match.group(1))
 
@@ -161,11 +160,18 @@ def check_file_exists_on_storage(run, tmpstorage):
 
     def go(name: str, path: str, size: int):
         path = tmpstorage + path
-        captured = run(["storage", "ls", path])
-        captured_output_list = captured.out.split("\n")
-        expected_line = format_list(type="file", size=size, name=name)
+        captured = run(["storage", "ls", "-l", path])
         assert not captured.err
-        assert expected_line in captured_output_list
+        files = output_to_files(captured.out)
+        for file in files:
+            if (
+                file.type == FileStatusType.FILE
+                and file.name == name
+                and file.size == size
+            ):
+                break
+        else:
+            raise AssertionError(f"File {name} with size {size} not found in {path}")
 
     return go
 
@@ -179,10 +185,14 @@ def check_dir_exists_on_storage(run, tmpstorage):
 
     def go(name: str, path: str):
         path = tmpstorage + path
-        captured = run(["storage", "ls", path])
-        captured_output_list = captured.out.split("\n")
-        assert f"directory      0              {name}" in captured_output_list
+        captured = run(["storage", "ls", "-l", path])
         assert not captured.err
+        files = output_to_files(captured.out)
+        for file in files:
+            if file.type == FileStatusType.DIRECTORY and file.path == name:
+                break
+        else:
+            raise AssertionError(f"Dir {name} not found in {path}")
 
     return go
 
@@ -196,10 +206,12 @@ def check_dir_absent_on_storage(run, tmpstorage):
 
     def go(name: str, path: str):
         path = tmpstorage + path
-        captured = run(["storage", "ls", path])
-        split = captured.out.split("\n")
-        assert format_list(name=name, size=0, type="directory") not in split
+        captured = run(["storage", "ls", "-l", path])
         assert not captured.err
+        files = output_to_files(captured.out)
+        for file in files:
+            if file.type == FileStatusType.DIRECTORY and file.path == name:
+                raise AssertionError(f"Dir {name} found in {path}")
 
     return go
 
@@ -213,10 +225,12 @@ def check_file_absent_on_storage(run, tmpstorage):
 
     def go(name: str, path: str):
         path = tmpstorage + path
-        captured = run(["storage", "ls", path])
-        pattern = format_list_pattern(name=name)
-        assert not re.search(pattern, captured.out)
+        captured = run(["storage", "ls", "-l", path])
         assert not captured.err
+        files = output_to_files(captured.out)
+        for file in files:
+            if file.type == FileStatusType.FILE and file.path == name:
+                raise AssertionError(f"File {name} found in {path}")
 
     return go