+++ python/tests/e2e/test_e2e_jobs.py
@@ -606,7 +606,7 @@ def test_e2e_multiple_env_from_file(run, tmp_path):
 
 @pytest.mark.e2e
 def test_e2e_ssh_exec_true(run):
-    command = 'bash -c "sleep 1m; false"'
+    command = 'bash -c "sleep 15m; false"'
     captured = run(
         [
             "job",
@@ -632,7 +632,7 @@ def test_e2e_ssh_exec_true(run):
 
 @pytest.mark.e2e
 def test_e2e_ssh_exec_false(run):
-    command = 'bash -c "sleep 1m; false"'
+    command = 'bash -c "sleep 15m; false"'
     captured = run(
         [
             "job",
@@ -659,7 +659,7 @@ def test_e2e_ssh_exec_false(run):
 
 @pytest.mark.e2e
 def test_e2e_ssh_exec_no_cmd(run):
-    command = 'bash -c "sleep 1m; false"'
+    command = 'bash -c "sleep 15m; false"'
     captured = run(
         [
             "job",
@@ -686,7 +686,7 @@ def test_e2e_ssh_exec_no_cmd(run):
 
 @pytest.mark.e2e
 def test_e2e_ssh_exec_echo(run):
-    command = 'bash -c "sleep 1m; false"'
+    command = 'bash -c "sleep 15m; false"'
     captured = run(
         [
             "job",
@@ -712,7 +712,7 @@ def test_e2e_ssh_exec_echo(run):
 
 @pytest.mark.e2e
 def test_e2e_ssh_exec_no_tty(run):
-    command = 'bash -c "sleep 1m; false"'
+    command = 'bash -c "sleep 15m; false"'
     captured = run(
         [
             "job",
@@ -739,7 +739,7 @@ def test_e2e_ssh_exec_no_tty(run):
 
 @pytest.mark.e2e
 def test_e2e_ssh_exec_tty(run):
-    command = 'bash -c "sleep 1m; false"'
+    command = 'bash -c "sleep 15m; false"'
     captured = run(
         [
             "job",
@@ -796,3 +796,61 @@ def test_e2e_ssh_exec_dead_job(run):
     with pytest.raises(SystemExit) as cm:
         run(["job", "exec", "--no-key-check", job_id, "true"])
     assert cm.value.code == 127
+
+
+@pytest.mark.e2e
+def test_e2e_job_list_filtered_by_status(run):
+    N_JOBS = 5
+
+    # submit N jobs
+    jobs = set()
+    for _ in range(N_JOBS):
+        command = "sleep 10m"
+        captured = run(["job", "submit", UBUNTU_IMAGE_NAME, command, "--quiet"])
+        job_id = captured.out.strip()
+        wait_job_change_state_from(run, job_id, Status.PENDING)
+        jobs.add(job_id)
+
+    # no status filtering (same as running+pending)
+    captured = run(["job", "ls", "--quiet"])
+    out = captured.out.strip()
+    jobs_ls_no_arg = set(out.split("\n"))
+    # check '<=' (not '==') multiple builds run in parallel can interfere
+    assert jobs <= jobs_ls_no_arg
+
+    # 1 status filter: running
+    captured = run(["job", "ls", "--status", "running", "--quiet"])
+    out = captured.out.strip()
+    jobs_ls_running = set(out.split("\n"))
+    # check '<=' (not '==') multiple builds run in parallel can interfere
+    assert jobs <= jobs_ls_running
+
+    # 2 status filters: pending+running is the same as without arguments
+    captured = run(["job", "ls", "-s", "pending", "-s", "running", "-q"])
+    out = captured.out.strip()
+    jobs_ls_running = set(out.split("\n"))
+    # check '<=' (not '==') multiple builds run in parallel can interfere
+    assert jobs_ls_running <= jobs_ls_no_arg
+
+    # "all" status filter is the same as "running+pending+failed+succeeded"
+    captured = run(["job", "ls", "-s", "all", "-q"])
+    out = captured.out.strip()
+    jobs_ls_all = set(out.split("\n"))
+    captured = run(
+        [
+            "job",
+            "ls",
+            "-s",
+            "running",
+            "-s",
+            "pending",
+            "-s",
+            "failed",
+            "-s",
+            "succeeded",
+            "-q",
+        ]
+    )
+    out = captured.out.strip()
+    jobs_ls_all_explicit = set(out.split("\n"))
+    assert jobs_ls_all <= jobs_ls_all_explicit