+++ python/tests/e2e/utils.py
@@ -1,6 +1,10 @@
+import re
+import time
 from time import sleep
+from typing import Sequence
 
 from _sha1 import sha1
+from neuromation.client import FileStatus, FileStatusType
 
 
 BLOCK_SIZE_MB = 16
@@ -11,6 +15,41 @@ RC_TEXT = "url: https://platform.dev.neuromation.io/api/v1\nauth: {token}"
 UBUNTU_IMAGE_NAME = "ubuntu:latest"
 format_list = "{type:<15}{size:<15,}{name:<}".format
 format_list_pattern = "(file|directory)\\s*\\d+\\s*{name}".format
+
+file_format_re = (
+    r"(?P<type>[-d])"
+    r"(?P<permission>[rwm])\s+"
+    r"(?P<size>\d+)\s+"
+    r"(?P<time>\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2})\s+"
+    r"(?P<name>.+)"
+)
+
+
+def output_to_files(output: str) -> Sequence[FileStatus]:
+    result = []
+    for match in re.finditer(file_format_re, output):
+        type = FileStatusType.FILE
+        if match["type"] == "d":
+            type = FileStatusType.DIRECTORY
+
+        permission = "read"
+        if match["permission"] == "w":
+            permission = "write"
+        elif match["permission"] == "m":
+            permission = "manage"
+
+        ts = int(time.mktime(time.strptime(match["time"], "%Y-%m-%d %H:%M:%S")))
+
+        result.append(
+            FileStatus(
+                path=match["name"],
+                size=int(match["size"]),
+                type=type,
+                modification_time=ts,
+                permission=permission,
+            )
+        )
+    return result
 
 
 def hash_hex(file):